feat(fc): add fc resolver nodes from filesystem
