--- connector/extensions/catalog-node-resolver-fs/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.oaebudt.edc.fc.ParticipantsResolverExtension
