+++ connector/extensions/catalog-node-resolver/build.gradle.kts
@@ -3,10 +3,10 @@ plugins {
 }
 
 dependencies {
-    runtimeOnly(libs.edc.fc.core)
-    runtimeOnly(libs.edc.fc.ext.api)
-    implementation(libs.edc.fc.spi.crawler)
     implementation(libs.edc.spi.identity.did)
+    implementation(libs.edc.fc.spi.crawler)
+    runtimeOnly(libs.edc.fc.core)
+    runtimeOnly(libs.edc.fc.api)
 
     testImplementation(libs.edc.junit)
     testImplementation(libs.junit.jupiter)
