+++ connector/extensions/catalog-node-resolver/src/main/java/org/oaebudt/edc/fc/CatalogNodeDirectoryExtension.java
@@ -1,30 +1,66 @@
 package org.oaebudt.edc.fc;
 
 import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
-import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
+import org.eclipse.edc.crawler.spi.TargetNodeFilter;
+import org.eclipse.edc.iam.did.spi.resolution.DidResolverRegistry;
+import org.eclipse.edc.runtime.metamodel.annotation.Extension;
+import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provider;
-import org.eclipse.edc.runtime.metamodel.annotation.Setting;
-import org.eclipse.edc.runtime.metamodel.annotation.Settings;
+import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.system.ServiceExtension;
+import org.eclipse.edc.spi.system.ServiceExtensionContext;
+import org.eclipse.edc.spi.types.TypeManager;
 
-import java.util.Arrays;
+import java.io.File;
 
+@Extension(value = CatalogNodeDirectoryExtension.NAME)
 public class CatalogNodeDirectoryExtension implements ServiceExtension {
+    public static final String NAME = "Participant Resolver Extension";
 
-    @Configuration
-    private FcParticipantsListConfiguration fcParticipantsListConfiguration;
+    public static final String PARTICIPANT_LIST_FILE_PATH = "oaebudt.ds.participants.list.file";
 
-    @Provider 
-    public TargetNodeDirectory federatedCacheNodeDirectory() {
-        return new CatalogNodeDirectory(Arrays.asList(fcParticipantsListConfiguration.participants().split(",")));
+    @Inject
+    private TypeManager typeManager;
+
+    @Inject
+    private DidResolverRegistry didResolverRegistry;
+
+    private File participantListFile;
+    private Monitor monitor;
+    private TargetNodeDirectory nodeDirectory;
+
+    @Override
+    public String name() {
+        return NAME;
     }
 
-    @Settings
-    record FcParticipantsListConfiguration(
-            @Setting(key = "fc.participants.list", description = "List of participants to crawl", defaultValue = "no:op")
-            String participants
-    ) {
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        var participantsPath = context.getConfig().getString(PARTICIPANT_LIST_FILE_PATH);
+        monitor = context.getMonitor();
 
+        participantListFile = new File(participantsPath).getAbsoluteFile();
+        if (!participantListFile.exists()) {
+            monitor.warning("Path '%s' does not exist. It must be a resolvable path with read access. Will not add any VCs.".formatted(participantsPath));
+        }
     }
 
+    @Provider
+    public TargetNodeDirectory createLazyTargetNodeDirectory() {
+        if (nodeDirectory == null) {
+            nodeDirectory = new CatalogNodeDirectory(typeManager.getMapper(), participantListFile, didResolverRegistry, monitor);
+        }
+        return nodeDirectory;
+    }
+
+    @Provider
+    public TargetNodeFilter skipSelfNodeFilter(ServiceExtensionContext context) {
+        return targetNode -> {
+            var predicateTest = !targetNode.id().equals(context.getParticipantId());
+            if (!predicateTest) {
+                monitor.debug("Node filter: skipping node '%s' for participant '%s'".formatted(targetNode.id(), context.getParticipantId()));
+            }
+            return predicateTest;
+        };
+    }
 }
