+++ connector/settings.gradle.kts
@@ -24,7 +24,6 @@ include(":extensions:superuser-seed")
 include(":extensions:dcp-impl")
 include(":extensions:catalog-node-resolver")
 include(":extensions:credential-from-file")
-include(":extensions:catalog-node-resolver-fs")
 
 include(":launchers:controlplane")
 include(":launchers:dataplane")
